% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{as.list.tree_bag}
\alias{as.list.tree_bag}
\title{Translate a tree bag into an R list.}
\usage{
\method{as.list}{tree_bag}(x, ...)
}
\arguments{
\item{x}{The tree bag}

\item{...}{Additional parameters. Not used but part of the generic interface.}
}
\value{
An R list with the elements found in the linked list.
}
\description{
Translate a tree bag into an R list.
}

