% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{as.vector.tree_bag}
\alias{as.vector.tree_bag}
\title{Translate a tree bag into an R vector.}
\usage{
\method{as.vector}{tree_bag}(x, mode = "any")
}
\arguments{
\item{x}{The tree bag.}

\item{mode}{The mode (numeric, logical, etc.) for the constructed vector.
The default will be a list}
}
\value{
An R vector with the elements found in the bag.
}
\description{
Translate a tree bag into an R vector.
}

