% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{bag_cons}
\alias{bag_cons}
\title{Add a head item to a linked list implementation of a bag.}
\usage{
bag_cons(elem, lst)
}
\arguments{
\item{elem}{The item to put at the head of the bag}

\item{bag}{The bag -- it will become the tail of the new list.}
}
\value{
a new bag.
}
\description{
Add a head item to a linked list implementation of a bag.
}
