% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{bag_node}
\alias{bag_node}
\title{Add a head item to a tree implementation of a bag.}
\usage{
bag_node(elem, left, right)
}
\arguments{
\item{elem}{The item to put at the head of the bag}

\item{bag}{The bag -- it will become the tail of the new list.}
}
\value{
a new bag.
}
\description{
Add a head item to a tree implementation of a bag.
}
