% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaps.R
\name{delete_minimal}
\alias{delete_minimal}
\title{Delete the minimal value in a heap}
\usage{
delete_minimal(heap)
}
\arguments{
\item{heap}{The heap}
}
\value{
The heap with the minimal value removed
}
\description{
Delete the minimal value in a heap
}

