% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deques.R
\name{dequeue_back}
\alias{dequeue_back}
\title{Remove the back element of a deque}
\usage{
dequeue_back(x)
}
\arguments{
\item{x}{The deque}
}
\value{
The updated deque
}
\description{
Remove the back element of a deque
}

