% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deques.R
\name{dequeue_front}
\alias{dequeue_front}
\title{Remove the front element of a deque}
\usage{
dequeue_front(x)
}
\arguments{
\item{x}{The deque}
}
\value{
The updated deque
}
\description{
Remove the front element of a deque
}
