% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstras algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{is a data.frame contaning}

\item{init_node}{is a numeric value that exist in the graph data.frame}
}
\description{
Dijkstras algorithm
}
\details{
The algoritm takes a graph and an initial node and iteratively calculates the shortest distance between the initial node and every other node in the graph.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}

