% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queues.R
\name{empty_closure_queue}
\alias{empty_closure_queue}
\title{Construct an empty closure based queue}
\usage{
empty_closure_queue()
}
\value{
an empty queue
}
\description{
Construct an empty closure based queue
}
