% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queues.R
\name{empty_extended_queue}
\alias{empty_extended_queue}
\title{Construct an empty extended queue}
\usage{
empty_extended_queue()
}
\value{
an empty queue
}
\description{
This is just a queue that doesn't use a closure to be able to update
the data structure when front is called.
}
