% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_queue.R
\name{empty_rebuild_queue}
\alias{empty_rebuild_queue}
\title{Create an empty "global rebuild" queue}
\usage{
empty_rebuild_queue()
}
\value{
empty quuee
}
\description{
Create an empty "global rebuild" queue
}
