% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_trees.R
\name{empty_red_black_tree}
\alias{empty_red_black_tree}
\title{Create empty red-black search tree}
\usage{
empty_red_black_tree()
}
\value{
New, empty, red-black search tree
}
\description{
Create empty red-black search tree
}
