% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_trees.R
\name{empty_search_tree}
\alias{empty_search_tree}
\title{Create empty unbalanced search tree}
\usage{
empty_search_tree()
}
\value{
New empty, unbalanced, search tree
}
\description{
Create empty unbalanced search tree
}

