% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queues.R
\name{enqueue}
\alias{enqueue}
\title{Add an element to a queue}
\usage{
enqueue(x, elm)
}
\arguments{
\item{x}{A queue}

\item{elm}{An element}
}
\value{
an updated queue where the element has been added
}
\description{
Add an element to a queue
}
