% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deques.R
\name{enqueue_back}
\alias{enqueue_back}
\title{Add an element to the front of a deque}
\usage{
enqueue_back(x, elm)
}
\arguments{
\item{x}{A deque}

\item{elm}{An element}
}
\value{
an updated deque where the element has been added
}
\description{
Add an element to the front of a deque
}

