% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaps.R
\name{find_minimal}
\alias{find_minimal}
\title{Get the minimal value in a heap}
\usage{
find_minimal(heap)
}
\arguments{
\item{heap}{The heap}
}
\value{
the minimal value in the heap
}
\description{
Get the minimal value in a heap
}

