% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{insert}
\alias{insert}
\title{Insert a value into a data structure.}
\usage{
insert(x, elm, ...)
}
\arguments{
\item{x}{The data structure}

\item{elm}{The element}

\item{...}{Potential optional arguments}
}
\value{
The updated data structure
}
\description{
Insert a value into a data structure.
}
