% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaps.R
\name{is_empty.binomial_heap}
\alias{is_empty.binomial_heap}
\title{Test whether a binomial heap is empty}
\usage{
\method{is_empty}{binomial_heap}(x)
}
\arguments{
\item{x}{binomial heap}
}
\value{
Whether the heap is empty
}
\description{
Test whether a binomial heap is empty
}
