% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaps.R
\name{is_empty.leftist_heap}
\alias{is_empty.leftist_heap}
\title{Test whether a leftist heap is empty}
\usage{
\method{is_empty}{leftist_heap}(x)
}
\arguments{
\item{x}{Leftist heap}
}
\value{
Whether the heap is empty
}
\description{
Test whether a leftist heap is empty
}

