% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{is_empty}
\alias{is_empty}
\title{Test if a data structure is empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{The data structure}
}
\value{
TRUE if x is empty.
}
\description{
Test if a data structure is empty
}

