% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_trees.R
\name{is_empty.red_black_tree}
\alias{is_empty.red_black_tree}
\title{Check if a red-black search tree is empty}
\usage{
\method{is_empty}{red_black_tree}(x)
}
\description{
Check if a red-black search tree is empty
}

