% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaps.R
\name{is_empty.splay_heap}
\alias{is_empty.splay_heap}
\title{Test whether a splay heap is empty}
\usage{
\method{is_empty}{splay_heap}(x)
}
\arguments{
\item{x}{splay heap}
}
\value{
Whether the heap is empty
}
\description{
Test whether a splay heap is empty
}

