% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_trees.R
\name{is_empty.unbalanced_search_tree}
\alias{is_empty.unbalanced_search_tree}
\title{Check if an unbalanced search tree is empty}
\usage{
\method{is_empty}{unbalanced_search_tree}(x)
}
\description{
Check if an unbalanced search tree is empty
}
