% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{list_concat_head}
\alias{list_concat_head}
\title{Add a head item to a linked list.}
\usage{
list_concat_head(lst, elem)
}
\arguments{
\item{lst}{The list -- it will become the tail of the new list.}

\item{elem}{The item to put at the head of the list.}
}
\value{
a new linked list.
}
\description{
Add a head item to a linked list.
}

