% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{list_drop_n}
\alias{list_drop_n}
\title{Get a new list containing the elements when n elements have been removed from a list/}
\usage{
list_drop_n(lst, n)
}
\arguments{
\item{lst}{A list}

\item{n}{the number of elements to remove}
}
\value{
a new list containing the elements of lst except the first n
}
\description{
Get a new list containing the elements when n elements have been removed from a list/
}
