% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{list_get_n}
\alias{list_get_n}
\title{Get a new list containing the first n elements of a list.}
\usage{
list_get_n(lst, n)
}
\arguments{
\item{lst}{A list}

\item{n}{the number of elements to get}
}
\value{
a new list containing the first n elements of lst.
}
\description{
Get a new list containing the first n elements of a list.
}
