% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{list_head}
\alias{list_head}
\title{Get the item at the head of a linked list.}
\usage{
list_head(lst)
}
\arguments{
\item{lst}{The list}
}
\value{
The element at the head of the list.
}
\description{
Get the item at the head of a linked list.
}

