% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{list_length}
\alias{list_length}
\title{Compute the length of a list.}
\usage{
list_length(lst)
}
\arguments{
\item{lst}{The lsit}
}
\value{
The number of elements in the list.
}
\description{
This is a linear time algorithm.
}
