% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{member}
\alias{member}
\title{Check membership in a data structure}
\usage{
member(x, elm, ...)
}
\arguments{
\item{x}{The data structure}

\item{elm}{Element to check membership for}

\item{...}{Potential optional arguments}
}
\value{
the merged x and y
}
\description{
Check membership in a data structure
}

