% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{merge}
\alias{merge}
\title{Merge two data structures (of the same kind)}
\usage{
merge(x, y, ...)
}
\arguments{
\item{x}{The first data structure}

\item{y}{The second data structure}

\item{...}{Potential optional arguments}
}
\value{
the merged x and y
}
\description{
Merge two data structures (of the same kind)
}

