% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{push}
\alias{push}
\title{Add an element to the top of a stack.}
\usage{
push(x, elm)
}
\arguments{
\item{x}{The stack}

\item{elm}{The elememtn}
}
\value{
an updated stack
}
\description{
Add an element to the top of a stack.
}
