% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_access_lists.R
\name{ral_head}
\alias{ral_head}
\title{Get the front element of a random access list}
\usage{
ral_head(ral)
}
\arguments{
\item{ral}{Random access list}
}
\value{
The front element
}
\description{
Get the front element of a random access list
}
