% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_access_lists.R
\name{ral_is_empty}
\alias{ral_is_empty}
\title{Check if a random access list is empty}
\usage{
ral_is_empty(ral)
}
\arguments{
\item{ral}{Random access list}
}
\value{
whether ral is empty
}
\description{
Check if a random access list is empty
}
