% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_access_lists.R
\name{ral_lookup}
\alias{ral_lookup}
\title{Get the value at a given index of a random access list.}
\usage{
ral_lookup(ral, idx)
}
\arguments{
\item{ral}{The random access list}

\item{idx}{The index to access}
}
\value{
The value at index idx
}
\description{
Get the value at a given index of a random access list.
}

