% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_access_lists.R
\name{ral_tail}
\alias{ral_tail}
\title{Get the tail of a random access list}
\usage{
ral_tail(ral)
}
\arguments{
\item{ral}{Random access list}
}
\value{
The tail of the list
}
\description{
Get the tail of a random access list
}

