% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_access_lists.R
\name{ral_update}
\alias{ral_update}
\title{Update the value at a given index of a random access list.}
\usage{
ral_update(ral, idx, value)
}
\arguments{
\item{ral}{The random access list}

\item{idx}{The index to access}

\item{value}{The new value to put into the list}
}
\value{
The value at index idx
}
\description{
Update the value at a given index of a random access list.
}
