% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{remove}
\alias{remove}
\title{Remove an element from a data structure}
\usage{
remove(x, elm, ...)
}
\arguments{
\item{x}{The data structure}

\item{elm}{Element to remove}

\item{...}{Potential optional arguments}
}
\value{
x with elm removed
}
\description{
Remove an element from a data structure
}

