% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{size}
\alias{size}
\title{Get the size of a data structure}
\usage{
size(x)
}
\arguments{
\item{x}{The data structure}
}
\value{
The number of elements in the data structure
}
\description{
Get the size of a data structure
}
