% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaps.R
\name{vector_to_heap}
\alias{vector_to_heap}
\title{Construct a heap from a vector of elements.}
\usage{
vector_to_heap(vec, empty_heap, empty_queue = empty_env_queue())
}
\arguments{
\item{vec}{The vector to translate into a heap.}

\item{empty_heap}{Empty heap for determing which heap implementation to use.}

\item{empty_queue}{Empty queue to determing which queue implementation to use for
building the heap.}
}
\description{
This function builds heaps in linear time by iteratively merging larger and larger heaps.
}
