% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Discription about the dataset included in the package'}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
wiki_graph
}
\arguments{
\item{v1}{is a vector of integers}

\item{v2}{is a vector of integers}

\item{w}{is a vector that contains weights of the edges}
}
\description{
Discription about the dataset included in the package'
}
\details{
This dataset contains a graph with tre varibles
and should be used as an example with the Dijkstra function
}
\author{
Yumeng Li, Mattias Karlsson, Ashraf Sarhan
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm#Pseudocode}
}
\keyword{datasets}

