% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{list_length}
\alias{list_length}
\title{Compute the length of a linked list}
\usage{
list_length(lst)
}
\arguments{
\item{lst}{The list}
}
\value{
the length of the list
}
\description{
Compute the length of a linked list
}
