% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A dataset containing three variables, v1, v2 and w.}
\format{A data frame with 18 rows
\describe{
 \item{v1}{Contains edges for the graph}
 \item{v2}{Contains edges for the graph}
 \item{w}{Coresponds to the weight for each node}
}}
\usage{
wiki_graph
}
\description{
A dataset containing three variables, v1, v2 and w.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm#Pseudocode}
}
\keyword{datasets}

